refactor(tests): consolidate Python U-238 fixtures + fix 6.67/6.674 drift
Wave 2 PR-4b of the architecture-audit refactor sprint.  Closes
test-duplication v2 finding F13.  PR-4a (#584) shipped the Rust side;
this PR completes Wave 2 with the Python side.

Per .research/audit-r4-architecture-v2/SUMMARY-v2.md.

Primary value: **drift fix**, not LOC reduction.  Python
`_make_single_resonance` defaulted to `energy=6.67`; Rust
`u238_single_resonance()` uses `energy=6.674`.  Both intended to encode
the same U-238 6.674 eV resonance, but the 2-vs-3-sig-fig drift was
historically invisible because peak-position assertions tolerate a
4 meV shift.  The audit explicitly flagged this as the highest-risk
F13 sub-finding.  Drift is now eliminated by aligning Python defaults
to 6.674 in a single shared fixture module.

Sites consolidated (3 fixtures, total 44 LOC of in-test helpers):

  Old location                       → New shared module
  ----                                 ----
  tests/test_nereids.py:20-41        → tests/_fixtures.py::_make_single_resonance
  tests/test_mcp.py:290-302          → tests/_fixtures.py::_synthetic_u238_entry
  tests/test_mcp.py:305-313          → tests/_fixtures.py::_synthetic_u238_data

Energy alignment (6.67 → 6.674) applied in all three helpers.  Other
audit-out-of-scope defaults preserved verbatim: scattering_radius=9.48,
awr=236.006, j=0.5, gn=0.0015, gg=0.023, target_spin=0.0 (per user
decision to keep F13 scope to the energy drift only).

Imports added in `test_nereids.py` (1 LOC) and `test_mcp.py` (1 LOC).

LOC delta:
- inserted: +65
  - +60: tests/_fixtures.py (3 fixtures + 6-line module docstring + import)
  - +2:  one `from _fixtures import ...` line per test module
  - +3:  test_nereids.py `u238_data` docstring updated 6.67→6.674
- deleted:  -51
  - -22: tests/test_nereids.py `_make_single_resonance` helper body
  - -29: tests/test_mcp.py `_synthetic_u238_entry` + `_synthetic_u238_data` bodies
- net:      +14

Audit projected NET -3.  Engineering reality: a 3-helper consolidation
into a new shared Python module cannot beat ~+10 LOC of module-level
overhead (docstring, import statement, inter-fn blank lines).  The
audit's -3 projection assumed a 39-LOC fixture file; the actual minimum
clear-style file is ~60 LOC (cramming Python kwargs onto one line would
get to ~45 but at material readability cost).  This is documented as a
calibration data point for audit-projection accuracy on Python
consolidations: the audit's LOC arithmetic underestimated by ~17 LOC.

The +14 is the **cost of the drift fix's single-source-of-truth**.
Future helper consumers in this directory will now import from
`_fixtures.py` rather than re-declaring; the drift cannot re-emerge
without an explicit edit to the canonical default.

Existing-logic check:
- No new helpers beyond the 3 moved.  Module-level `import nereids` +
  6-line docstring are inherent to having any new Python module.
- No production code changes.  Pure test-side refactor + one numerical
  value alignment (6.67 → 6.674) per the audit's drift-fix mandate.
- No new helpers without ≥1 deletion: the 3 helpers in `_fixtures.py`
  are direct moves of 3 helpers from the test files.

Test plan:
- pixi run test-python: **164 passed, 0 failed** (unchanged from main).
  The 4 mEV shift (6.67 → 6.674) is well below all Python assertion
  tolerances on peak-position / cross-section magnitude — confirmed by
  zero test failures.

Linked: refactor sprint Wave 2 PR-4b per
.research/audit-r4-architecture-v2/SUMMARY-v2.md
